requirements.txt
python-dotenv
langgraph
langchain-google-vertexai
langchain-core
google-cloud-texttospeech
google-cloud-aiplatform
requests
google-api-python-client
youtube-transcript-api
pytest
PyGithub
tavily-python
GitPython
